requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
python-docx>=1.0.0
kaleido>=0.2.1
numpy>=1.24.0
scikit-learn>=1.3.0  # For semantic search
sentence-transformers>=2.2.2  # For semantic search